Refactor file structure to be more workable
